move sync_sources system to source submodule
