+++ audionimbus/src/bevy/runner/mod.rs
@@ -1,12 +1,8 @@
 //! Runner typestate: marker types that map simulation modes to dedicated threads.
 
 use super::configuration::SimulationConfiguration;
-use super::simulation::Simulation;
-use super::source::{Listener, Source, SourceParameters};
 use crate::sealed::Sealed;
-use crate::simulation::{SimulationInputs, SimulationParameters};
-use crate::wiring::SourceWithInputs;
-use bevy::prelude::{App, Entity, Query, Res, Transform, Without, World};
+use bevy::prelude::{App, World};
 
 mod direct;
 mod pathing;
@@ -55,30 +51,4 @@ pub trait SyncFrame<C: SimulationConfiguration>: Sealed {
 
 impl<C: SimulationConfiguration> SyncFrame<C> for () {
     fn add_systems(_app: &mut App) {}
-}
-
-pub(crate) fn sync_sources<C: SimulationConfiguration>(
-    mut query: Query<
-        (Entity, &Transform, &Source<C>, Option<&SourceParameters<C>>),
-        Without<Listener>,
-    >,
-    simulation: Res<Simulation<C>>,
-) {
-    simulation.0.update_sources(|snapshot| {
-        for (entity, transform, source, simulation_parameters) in query.iter_mut() {
-            let simulation_inputs = SimulationInputs {
-                source: (*transform).into(),
-                parameters: simulation_parameters
-                    .map_or_else(SimulationParameters::default, |params| params.0.clone()),
-            };
-
-            snapshot.push((
-                entity,
-                SourceWithInputs {
-                    source: source.0.clone(),
-                    simulation_inputs,
-                },
-            ));
-        }
-    });
 }
